Generate passwords and TLS certificates at provisioning time (#359)
