+++ marketplace/dev/Dockerfile
@@ -9,6 +9,7 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
         jq \
         make \
         python \
+        python-openssl \
         python-pip \
         python-setuptools \
         python-yaml \
